requirements.txt
fastapi
uvicorn
numpy
openvino
openvino-genai
Pillow
huggingface_hub
python-dotenv
mcp
requests
